+++ tests/test_context.py
@@ -193,6 +193,7 @@ def test_close_before_pop(runner):
     @click.pass_context
     def cli(ctx):
         ctx.obj = 'test'
+
         @ctx.call_on_close
         def foo():
             assert click.get_current_context().obj == 'test'
@@ -239,3 +240,19 @@ def test_make_pass_decorator_args(runner):
     result = runner.invoke(cli, ['test2'])
     assert not result.exception
     assert result.output == 'foocmd\n'
+
+
+def test_exit_not_standalone():
+    @click.command()
+    @click.pass_context
+    def cli(ctx):
+        ctx.exit(1)
+
+    assert cli.main([], 'test_exit_not_standalone', standalone_mode=False) == 1
+
+    @click.command()
+    @click.pass_context
+    def cli(ctx):
+        ctx.exit(0)
+
+    assert cli.main([], 'test_exit_not_standalone', standalone_mode=False) == 0